requirements.txt
python-dotenv>=1.0.0
python-telegram-bot>=21.0
openai>=1.0.0
firebase-admin>=6.0.0
requests>=2.28.0
